# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Set up a timer that you can share with others",
        "title": "BreakChrono",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/breaks": {
            "post": {
                "description": "create a new timer by passing the end date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new timer",
                "parameters": [
                    {
                        "description": "New break to add",
                        "name": "date",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Break"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PostResponse"
                        }
                    }
                }
            }
        },
        "/breaks/{id}": {
            "get": {
                "description": "get end date by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the end date",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Timer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Break"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Break": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                }
            }
        },
        "models.PostResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}